refactor: create dedicated classes for OpenSearch indexing and retrieval
