--- src/retrieval.py
@@ -1,1 +1,0 @@
-"""Module aimed at defining retrievers for RAG."""
